Add fallback in POSIXUtil when JNA fails
The way POSIXUtil initialized the JNA Native library in a static field
meant if the call failed, the class itself would fail to initialize. To
handle this case more gracefully, move the JNA call to a static
initializer which catches any exceptions.

Additionally, add a fallback to getHostName that shells out to
/bin/hostname when JNA fails.