refactor: avoid generating getters for "constants"
It is unlikely that the original intent was to expose getters such as
`getBUILT_BY_PROPERTY()`, so change those properties to public fields,
for the benefit of a clearer API, and more idiomatic integration with
statically compiled languages.

https://docs.groovy-lang.org/docs/groovy-3.0.13/html/documentation/#_fields_and_properties

That change is source-compatible with existing dynamically compiled
Groovy, but breaks compatibility with any statically compiled language
(including Groovy with @CompileStatic): as the getters are removed, they
now need to refer to the newly exposed fields.